Add audio system with manager, categories, and elevator SFX
Introduced a new audio system including AudioManager and FMODEvents prefabs and scripts, as well as audio category scripts for background music and sound effects. Added an Audio Emmiter prefab and integrated it into the ElevatorDoor prefab, replacing the previous event emitter. Updated the Player Interaction Prototype scene to include an elevator button UI and connected it to the platform logic. Refactored and expanded audio-related scripts to support new audio events and categories.